ci: Add bigtable cleanup script
